requirements.txt
boto==2.37.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-appconf==1.0.1
django-bootstrap-pagination==1.5.1
django-bootstrap3==5.2.0
django-etc==0.7.0
django-geoposition==0.2.2
django-sitegate==0.10.0
django-storages==1.1.8
django-toolbelt==0.0.1
django-widget-tweaks==1.3
djangoaddattr==0.0.1
gunicorn==19.3.0
Pillow==2.8.1
psycopg2==2.6
six==1.9.0
static3==0.5.1
